feat: created useRefreshToken & useAttachAuthToken & interceptors refactoring
