+++ src/entities/session/index.ts
@@ -1,1 +1,9 @@
 export { getMe, loginFn, logoutFn, registerFn } from './api';
+export {
+  selectAccessToken,
+  selectIsAuthenticated,
+  selectStatus,
+} from './model/auth.selectors';
+export { useAuthStore } from './model/auth.store';
+export type { AuthState, AuthStatus } from './model/auth.types';
+export { useAuthMeQuery } from './model/hooks';
