+++ src/entities/session/model/auth.selectors.ts
