+++ src/entities/session/model/auth.types.ts
