+++ src/entities/session/model/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useAuthMeQuery } from './useAuthMeQuery';
