--- src/features/auth/index.ts
@@ -1,11 +1,0 @@
-export {
-  selectAccessToken,
-  selectIsAuthenticated,
-  selectStatus,
-} from './model/auth.selectors';
-export { useAuthStore } from './model/auth.store';
-export type { AuthState, AuthStatus } from './model/auth.types';
-export { useAuthMeQuery } from './model/hooks/useAuthMeQuery';
-export { useLoginMutation } from './model/hooks/useLoginMutation';
-export { useLogoutMutation } from './model/hooks/useLogoutMutation';
-export { useRegisterMutation } from './model/hooks/useRegisterMutation';
