--- src/features/auth/model/hooks/index.ts
@@ -1,4 +1,0 @@
-export { useAuthMeQuery } from './useAuthMeQuery';
-export { useLoginMutation } from './useLoginMutation';
-export { useLogoutMutation } from './useLogoutMutation';
-export { useRegisterMutation } from './useRegisterMutation';
