+++ src/features/session/model/useRegisterMutation.ts
@@ -1,20 +1,21 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { useAuthStore } from '../auth.store';
-
-import { loginFn } from '@/entities/session';
+import { registerFn, useAuthStore } from '@/entities/session';
 import { useUserStore } from '@/entities/user';
 
-import type { LoginRequest, LoginResponse } from '@/shared/api/contracts/auth';
+import type {
+  RegisterRequest,
+  RegisterResponse,
+} from '@/shared/api/contracts/auth';
 
-export const useLoginMutation = () => {
+export const useRegisterMutation = () => {
   const queryClient = useQueryClient();
   const authActions = useAuthStore((s) => s.actions);
   const userActions = useUserStore((s) => s.actions);
 
-  return useMutation<LoginResponse, Error, LoginRequest>({
+  return useMutation<RegisterResponse, Error, RegisterRequest>({
     mutationFn: async (payload) => {
-      const response = await loginFn('/auth/login', { arg: payload });
+      const response = await registerFn('/auth/register', { arg: payload });
       return response.data;
     },
     onSuccess: (data) => {
